2_year_2_semester/ser_app_dev_tech/devops-learning-app/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.23.2
sqlalchemy==2.0.22
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2